+++ NEWS.md
@@ -1,3 +1,5 @@
+# bigrquery (development version)
+
 # bigrquery 1.1.0
 
 ## Improved type support