+++ NEWS.md
@@ -1,4 +1,6 @@
-# bigrquery (development version)
+# bigrquery 1.1.1
+
+* Fix test failure with dbplyr 1.4.0.
 
 * `bq_field()` can now pass `description` parameter which will be applied
   in `bq_table_create()` call (@byapparov, #272).